Haptics: add intensity support
As it turns out, the AW8624_REG_DRV_LVL register is responsible for regulating the intensity of the vibration motor in continuous mode. I assumed the AW8624_REG_DRV_LVL_OV value (0x9B) would be the maximum allowed intensity level, but perhaps it might be possible to go beyond that, yet I wouldn't want to risk anything here.